refactor(adaptive_card): split monolithic file into modular structure
Split the 1312-line adaptive_card.rs into a well-organized directory:
- mod.rs (342 lines) - Core AdaptiveCard struct and Action enum
- elements.rs (748 lines) - CardElement enum with all implementations
- containers.rs (110 lines) - Column, Choice, and Fact structs
- styles.rs (146 lines) - All style enums (Color, Spacing, etc.)

Benefits:
- Better code organization and maintainability
- Easier to navigate and understand
- Logical grouping of related types
- All tests passing (37/37)
